+++ src/lib/turnClient.js
@@ -1,4 +1,4 @@
-export { dispatchTurnActivity, dispatchTurnEvent } from './turnClient/turnEventDispatch.js'
+export { createBufferedTurnActivityDispatcher, dispatchTurnActivity, dispatchTurnEvent } from './turnClient/turnEventDispatch.js'
 export { runServerTurn } from './turnClient/runServerTurn.js'
 export { fetchServerSessionSnapshot, normalizeServerSessionSnapshot } from './turnClient/sessionSnapshot.js'
 export { cancelServerTurn, getServerTurn, replayServerTurn, resumeServerTurnRequest, startServerTurn, steerServerTurn } from './turnClient/turnRequests.js'